Allow a case of query() where /count is not called
I'm hitting what appears to be a PowerSchool bug with PowerQueries (not
an infrequent occurence) which causes the query itself to work fine,
but the corresponding /count endpoint to return a nonsensical error.
Changing query() to not call /count in the pagesize=0 case allows for
this to be worked around.